allow toggling help and subdomain sections
Previously the 'more info' link would only show the help text and there
was no way to toggle it on and off. This change allows the user to
toggle the info/help on and off. Additionally, the make subdomain form
is now separate from the more info section. This makes the help text
more readable as well as making the make subdomain link more usable.